Added Line Status for most lines (still sorting out H&C and Circle), added train's current location, edited menu
The Line Status is shown in an on-screen button with custom background
resource helping to state the line status (green good, orange not so).
Edited the custom title bar to be a part of the activity's layout.
Moved certain methods to UiElements. Simplified getting and parsing the
JSON requests' response. Edited other minor UI components to adjust for
location and line status button.