Added onClick for Line Status button, which now shows a dialog with more details. All lines now supported for this. Added class with static variables just to have standard codes and IDs documented.
The button now properly displays status information for Hammersmith &
City and Circle lines (which are grouped into one), and displays a
dialog with more information when a user clicks it. The methods written
are a little ungainly and long.
StandardCodes was added to provide a class filled with standard IDs and
codes, nothing else.