Making headway into the development of the Widget. Made UiController into an abstract class, UiControllerMain now the UI controlling class for WheresMyTrain's main activity.
Made and amended layouts to make the homescreen widget 4x1, with a configuration activity (which currently does not open, but will be activated soon). Already fetches and displays (using the toString method) the detailed predictions for a given line and station (hardcoded for now to Central line and Charing Cross station). Uses a service, which it enables and disables with the first/last widget being added to the screen.
Will be adding use of Alarms from Android SDK for updates next, meaning shorter time between updates, but only updates when the device is already awake. Balances out on battery usage and network activity that way.